refactor: factor link modal into own file
